added in support for installing cuda on opensuse and ubuntu
Added in sles support for GPU config

cleaned up some linting errors with ansible, added in sles support for rancher_docker, and got GPU working on sles

added in home directory variables and moved default rancher image tag to inventory file to allow for overwrite on a per host basis

updated readmes for parent and child playbooks. Added in toggle for running the gpu_test to the default.yml file.